+++ src/challenges/chess-simp.ts
@@ -12,7 +12,7 @@ import {
   Challenge_2022_05_30,
   Challenge_2022_05_31,
 } from './chess-simp/2022-05'
-import { _2022_06_03 } from './chess-simp/2022-06'
+import { Challenge_2022_06_03, Challenge_2022_06_23 } from './chess-simp/2022-06'
 import { Challenge_2022_07_11, Challenge_2022_07_18 } from './chess-simp/2022-07'
 import {
   Challenge_2022_09_11,
@@ -57,7 +57,10 @@ export const chessSimpChallenges: Map<Uuid, { meta: ChallengeMeta; create: () =>
         () => new Challenge_2023_05_23() as Challenge,
         () => new Challenge_2022_05_31() as Challenge,
       ],
-      [() => _2022_06_03],
+      [
+        () => new Challenge_2022_06_03() as Challenge,
+        () => new Challenge_2022_06_23() as Challenge,
+      ],
       [
         () => new Challenge_2022_07_18() as Challenge,
         () => new Challenge_2022_07_11() as Challenge,