+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Yet unannounced in Chess Simp Discord
 
-* 2023-08-17: New challenge: Chess Simp -- "Chess, But I'm A Real Simp"
+* Aug 17, 2023: New challenge: Chess Simp -- "Chess, But I'm A Real Simp"
+* Aug 17, 2023: New challenge: Chess Simp -- "Poor Pawn Got Blundered 10 Times"
 
 # 2023-08-17 - started the changelog